+++ functions/src/index.ts
@@ -41,7 +41,7 @@ export * from './transact'
 export * from './change-user-info'
 export * from './create-user'
 export * from './create-answer'
-export * from './place-bet'
+export * from '../../web/lib/api/place-bet'
 export * from './cancel-bet'
 export * from './sell-bet'
 export * from './sell-shares'
@@ -63,7 +63,7 @@ import { transact } from './transact'
 import { changeuserinfo } from './change-user-info'
 import { createuser } from './create-user'
 import { createanswer } from './create-answer'
-import { placebet } from './place-bet'
+import { placebet } from '../../web/lib/api/place-bet'
 import { cancelbet } from './cancel-bet'
 import { sellbet } from './sell-bet'
 import { sellshares } from './sell-shares'
