+++ functions/src/serve.ts
@@ -13,7 +13,7 @@ import { transact } from './transact'
 import { changeuserinfo } from './change-user-info'
 import { createuser } from './create-user'
 import { createanswer } from './create-answer'
-import { placebet } from './place-bet'
+import { placebet } from '../../web/lib/api/place-bet'
 import { cancelbet } from './cancel-bet'
 import { sellbet } from './sell-bet'
 import { sellshares } from './sell-shares'
